@@ -1,16 +1,16 @@
 <Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
     <IsPackable>false</IsPackable>
   </PropertyGroup>
   <ItemGroup>
     <ProjectReference Include="..\Blazorise.DataGrid.Template\Blazorise.DataGrid.Template.csproj"/>
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.9.4"/>
-    <PackageReference Include="MSTest.TestAdapter" Version="2.2.3"/>
-    <PackageReference Include="MSTest.TestFramework" Version="2.2.3"/>
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.11.0"/>
+    <PackageReference Include="MSTest.TestAdapter" Version="2.2.7"/>
+    <PackageReference Include="MSTest.TestFramework" Version="2.2.7"/>
     <PackageReference Include="Dapper" Version="2.0.90"/>
-    <PackageReference Include="System.Data.SqlClient" Version="4.8.3"/>
+    <PackageReference Include="Microsoft.Data.SqlClient" Version="3.0.0"/>
   </ItemGroup>
 </Project>
