@@ -1,13 +1,11 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
-
   <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
   </PropertyGroup>
-
   <ItemGroup>
-    <PackageReference Include="Blazorise.Bootstrap" Version="0.9.4.4" />
-    <PackageReference Include="Blazorise.DataGrid" Version="0.9.4.4" />
-    <PackageReference Include="Blazorise.Icons.FontAwesome" Version="0.9.4.4" />
+    <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="6.0.0-preview.5.21301.5"/>
+    <PackageReference Include="Blazorise.Bootstrap" Version="0.9.4.4"/>
+    <PackageReference Include="Blazorise.DataGrid" Version="0.9.4.4"/>
+    <PackageReference Include="Blazorise.Icons.FontAwesome" Version="0.9.4.4"/>
   </ItemGroup>
-
 </Project>
